+++ config/versiones.php
@@ -1,10 +1,10 @@
 <?php
 
 return [
-    'IEE'           => '1.17.0',
-    'BhagamApps'    => '1.16.2',
+    'IEE'           => '1.18.0',
+    'BhagamApps'    => '1.17.0',
     'User'          => '2.4.4',
-    'Inventario'    => '2.12.0',
+    'Inventario'    => '2.13.0',
     'Apps'          => '1.5.1',
     'CrudGenerator' => '1.1.0',
 ];